requirements.txt
torchvision
opencv-python
numpy
Pillow
visdom
matplotlib
thop
tqdm